refactor: use get in common utils
